sys/log; bring in log specific data structures from fs/fcb.
